get back to wp editor button working
